fix: 가격 double로 수정

+++ src/main/java/org/phoenix/planet/dto/eco_stock/raw/MemberStockInfo.java
@@ -7,6 +7,6 @@ public record MemberStockInfo(
         Long ecoStockId,
         Integer currentTotalQuantity,
         Double currentTotalAmount,
-        Integer point
+        Double point
 ) {
 }
